implement the updateFormField handler for the min/max input fields
i'd like the minRequired value to affect the min attribute of the Maximum allowed input. and vice versa maxAllowed should affect the max attribute of the Minimum required input. this is to help with validation.

also, when user enters non numeric values into the input, would :invalid css apply and make it look invalid?