In CollectData view, checkboxes with min or max set should not carry the `(Optional)` suffix in the label
In CollectData view, for checkboxes with min or max set and is currently invalid, we are showing validationMessage to the user. Can we NOT show the validationMessage text but instead just set a class on the checkboxes? And add that css class alongside .tff-field-group.tff-required select:invalid in @input.css ?

Update @checkbox-constraints.spec.ts if it was looking for the validation message